fix: переделана механика обработки запросов на сервер - определяется тип запроса, dto отправляется в соответсвующий контроллер

+++ modules/dto/src/main/java/io/deeplay/intership/dto/response/FinishGameDtoResponse.java
@@ -4,15 +4,15 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class FinishGameDtoResponse extends BaseDtoResponse {
-    public final int blackScore;
-    public final int whiteScore;
+    public final double blackScore;
+    public final double whiteScore;
 
     @JsonCreator
     public FinishGameDtoResponse(
             @JsonProperty("status") ResponseStatus status,
             @JsonProperty("message") String message,
-            @JsonProperty("blackScore") int blackScore,
-            @JsonProperty("whiteScore") int whiteScore) {
+            @JsonProperty("blackScore") double blackScore,
+            @JsonProperty("whiteScore") double whiteScore) {
         super(status, message);
         this.blackScore = blackScore;
         this.whiteScore = whiteScore;
